Update treble boosting filter
This is a filter that is designed to boost all the way up to Nyquist, unlike the previous one that rolled off at 15/16 of Nyquist. Noting https://github.com/Optiroc/BRRtools/issues/5 , the values have not been multiplied by 0.6, so filtered samples will be somewhat likely to clip and/or overflow, but they will not be inherently attenuated in any way.